requirements.txt
flask
python-dotenv
authlib
requests
bandit